Fix bug that allowed illegal moves.
If movement in a direction causes no change, the move should be thrown out.
I wasn't verifying that the attempted move caused change, and I was still
randomizing a new 2/4 onto the board. This change fixes that by checking to be
sure the new board (after the attempted move) is different than the old board.